muda o provedor do banco de dados para PostgreSQL e adiciona suporte a enums para status de pedidos e papéis de usuários; implementa novos métodos para buscar e gerenciar carrinhos e pedidos por usuário e administrador

+++ src/product/product.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from '../../prisma/prisma.service';
 import { Prisma } from '@prisma/client';
 
@@ -6,7 +10,19 @@ import { Prisma } from '@prisma/client';
 export class ProductService {
   constructor(private prisma: PrismaService) {}
 
-  async create(createProductDto: Prisma.ProductCreateInput) {
+  async create(createProductDto: Prisma.ProductCreateInput, adminId: string) {
+    const user = await this.prisma.user.findUnique({
+      where: { id: adminId },
+    });
+
+    if (!user) {
+      throw new NotFoundException(`Admin with ID ${adminId} not found`);
+    }
+
+    if (user.role !== 'ADMIN') {
+      throw new ForbiddenException(`User with ID ${adminId} is not an admin`);
+    }
+
     return await this.prisma.product.create({
       data: createProductDto,
     });
@@ -22,14 +38,39 @@ export class ProductService {
     });
   }
 
-  async update(id: string, updateProductDto: Prisma.ProductUpdateInput) {
+  async update(
+    adminId: string,
+    id: string,
+    updateProductDto: Prisma.ProductUpdateInput,
+  ) {
+    const user = await this.prisma.user.findUnique({
+      where: { id: adminId },
+    });
+    if (!user) {
+      throw new NotFoundException(`Admin with ID ${adminId} not found`);
+    }
+
+    if (user.role !== 'ADMIN') {
+      throw new ForbiddenException(`User with ID ${adminId} is not an admin`);
+    }
     return await this.prisma.product.update({
       where: { id },
       data: updateProductDto,
     });
   }
 
-  async remove(id: string) {
+  async remove(id: string, adminId: string) {
+    const user = await this.prisma.user.findUnique({
+      where: { id: adminId },
+    });
+    if (!user) {
+      throw new NotFoundException(`Admin with ID ${adminId} not found`);
+    }
+
+    if (user.role !== 'ADMIN') {
+      throw new ForbiddenException(`User with ID ${adminId} is not an admin`);
+    }
+
     return await this.prisma.product.delete({
       where: { id },
     });
